+++ src-docs/src/views/guidelines/sass.js
@@ -250,7 +250,7 @@ function renderAnimationTiming(speed) {
   );
 }
 
-function renderBreakpoint(size) {
+export function renderBreakpoint(size, breakpoints) {
   return (
     <EuiFlexGroup
       responsive={false}
@@ -593,7 +593,7 @@ export const SassGuidelines = ({ selectedTheme }) => {
 
       <GuideRuleTitle>Typography</GuideRuleTitle>
 
-      <EuiText grow={false} className="guideSection__text">
+      <EuiText grow={false}>
         <p>
           View the{' '}
           <EuiLink href="https://github.com/elastic/eui/blob/master/src/global_styling/variables/_typography.scss">
@@ -878,7 +878,7 @@ export const SassGuidelines = ({ selectedTheme }) => {
 
       <GuideRuleTitle>Media queries and breakpoints</GuideRuleTitle>
 
-      <EuiText className="guideSection__text">
+      <EuiText>
         <p>
           <EuiLink href="https://github.com/elastic/eui/blob/master/src/global_styling/mixins/_responsive.scss">
             View the Sass code for media queries
@@ -905,8 +905,8 @@ export const SassGuidelines = ({ selectedTheme }) => {
 
             <EuiSpacer />
 
-            {euiBreakPoints.map(function (size, index) {
-              return renderBreakpoint(size, index);
+            {euiBreakPoints.map(function (size) {
+              return renderBreakpoint(size, breakpoints);
             })}
           </div>
         </EuiFlexItem>
@@ -958,7 +958,7 @@ export const SassGuidelines = ({ selectedTheme }) => {
       <EuiSpacer size="xxl" />
 
       <GuideRuleTitle>Animation</GuideRuleTitle>
-      <EuiText grow={false} className="guideSection__text">
+      <EuiText grow={false}>
         <p>
           <EuiLink href="https://github.com/elastic/eui/blob/master/src/global_styling/variables/_animations.scss">
             View the Sass code for animation
@@ -1026,7 +1026,7 @@ export const SassGuidelines = ({ selectedTheme }) => {
           </EuiCodeBlock>
         </EuiFlexItem>
         <EuiFlexItem>
-          <EuiText grow={false} className="guideSection__text">
+          <EuiText grow={false}>
             <h3>Writing Sass the EUI way</h3>
             <p>
               In general, when writing new SCSS in a project that installs EUI
@@ -1034,7 +1034,7 @@ export const SassGuidelines = ({ selectedTheme }) => {
             </p>
           </EuiText>
           <EuiSpacer />
-          <EuiText size="s" grow={false} className="guideSection__text">
+          <EuiText size="s" grow={false}>
             <ul>
               <li>
                 Utilize color variables and functions rather than hard-coded
@@ -1073,7 +1073,7 @@ export const SassGuidelines = ({ selectedTheme }) => {
 
           <EuiSpacer />
 
-          <EuiText grow={false} className="guideSection__text">
+          <EuiText grow={false}>
             <p>
               Most EUI based projects should already import the EUI global
               scope. For example, Kibana has its own liner that will give you
@@ -1088,7 +1088,7 @@ export const SassGuidelines = ({ selectedTheme }) => {
             {importKibanaExample}
           </EuiCodeBlock>
           <EuiSpacer />
-          <EuiText grow={false} className="guideSection__text">
+          <EuiText grow={false}>
             <p>
               If you want to construct your own import, you would just need to
               import the following core files into a fresh Sass project.