+++ src/components/panel/panel.tsx
@@ -21,13 +21,11 @@ import React, {
   ButtonHTMLAttributes,
   FunctionComponent,
   HTMLAttributes,
-  ReactNode,
   Ref,
 } from 'react';
 import classNames from 'classnames';
 
 import { CommonProps, keysOf, ExclusiveUnion } from '../common';
-import { EuiBetaBadge } from '../badge/beta_badge';
 
 export const panelPaddingValues = {
   none: 0,
@@ -97,32 +95,21 @@ export interface _EuiPanelProps extends CommonProps {
    * Usually a lightened form of the brand colors
    */
   color?: PanelColor;
-  /**
-   * **DEPRECATED: use `EuiCard` instead.**
-   * Add a badge to the panel to label it as "Beta" or other non-GA state
-   */
-  betaBadgeLabel?: string;
-  /**
-   * **DEPRECATED: use `EuiCard` instead.**
-   * Add a description to the beta badge (will appear in a tooltip)
-   */
-  betaBadgeTooltipContent?: ReactNode;
-  /**
-   * **DEPRECATED: use `EuiCard` instead.**
-   * Optional title will be supplied as tooltip title or title attribute otherwise the label will be used
-   */
-  betaBadgeTitle?: string;
 }
 
 interface Divlike
   extends _EuiPanelProps,
-    Omit<HTMLAttributes<HTMLDivElement>, 'onClick' | 'color'> {}
+    Omit<HTMLAttributes<HTMLDivElement>, 'color'> {
+  element?: 'div';
+}
 
 interface Buttonlike
   extends _EuiPanelProps,
-    Omit<ButtonHTMLAttributes<HTMLButtonElement>, 'color'> {}
+    Omit<ButtonHTMLAttributes<HTMLButtonElement>, 'color'> {
+  element?: 'button';
+}
 
-export type EuiPanelProps = ExclusiveUnion<Divlike, Buttonlike>;
+export type EuiPanelProps = ExclusiveUnion<Buttonlike, Divlike>;
 
 export const EuiPanel: FunctionComponent<EuiPanelProps> = ({
   children,
@@ -134,10 +121,7 @@ export const EuiPanel: FunctionComponent<EuiPanelProps> = ({
   hasBorder,
   grow = true,
   panelRef,
-  onClick,
-  betaBadgeLabel,
-  betaBadgeTooltipContent,
-  betaBadgeTitle,
+  element,
   ...rest
 }) => {
   // Shadows are only allowed when there's a white background (plain)
@@ -157,34 +141,17 @@ export const EuiPanel: FunctionComponent<EuiPanelProps> = ({
       'euiPanel--hasBorder': canHaveBorder && hasBorder === true,
       'euiPanel--noBorder': !canHaveBorder || hasBorder === false,
       'euiPanel--flexGrowZero': !grow,
-      'euiPanel--isClickable': onClick,
-      'euiPanel--hasBetaBadge': betaBadgeLabel,
+      'euiPanel--isClickable': rest.onClick,
     },
     className
   );
 
-  let optionalBetaBadge;
-  if (betaBadgeLabel) {
-    optionalBetaBadge = (
-      <span className="euiPanel__betaBadgeWrapper">
-        <EuiBetaBadge
-          label={betaBadgeLabel}
-          title={betaBadgeTitle}
-          tooltipContent={betaBadgeTooltipContent}
-          className="euiPanel__betaBadge"
-        />
-      </span>
-    );
-  }
-
-  if (onClick) {
+  if (rest.onClick && element !== 'div') {
     return (
       <button
         ref={panelRef as Ref<HTMLButtonElement>}
         className={classes}
-        onClick={onClick}
         {...(rest as ButtonHTMLAttributes<HTMLButtonElement>)}>
-        {optionalBetaBadge}
         {children}
       </button>
     );
@@ -195,7 +162,6 @@ export const EuiPanel: FunctionComponent<EuiPanelProps> = ({
       ref={panelRef as Ref<HTMLDivElement>}
       className={classes}
       {...(rest as HTMLAttributes<HTMLDivElement>)}>
-      {optionalBetaBadge}
       {children}
     </div>
   );